==> 00_data_import.ipynb <==
**19/12/2021**
<br>
Bruno Mazzilli González

> # Import Data

import numpy as np
import pandas as pd
import warnings
warnings.filterwarnings("ignore")

#verificamos cuales son los valores desconocidos usando el documento "drivingLegend.pdf". En base a esto especificamos
#cuales son los valores na en nuestro csv:
cca_df = pd.read_csv('data/NCDB_1999_to_2014.csv'
                     ,na_values=["UU", "XX", "U", "X", "QQ", "Q", "NNNN", "UUUU", "XXXX", "NN", "N" ])

print(cca_df.shape)
cca_df.head()

cca_df.tail()

cca_df.columns

#check Python platform’s pointer that dictates the maximum size of lists and strings in Python
import sys
sys.maxsize == (2**63)-1

cca_df.dtypes

def dame_variables_categoricas(dataset=None):
    '''
    ----------------------------------------------------------------------------------------------------------
    Función dame_variables_categoricas:
    ----------------------------------------------------------------------------------------------------------
        -Descripción: Función que recibe un dataset y devuelve una lista con los nombres de las 
        variables categóricas
        -Inputs: 
            -- dataset: Pandas dataframe que contiene los datos
        -Return:
            -- lista_variables_categoricas: lista con los nombres de las variables categóricas del
            dataset de entrada con menos de 100 valores diferentes
            -- 1: la ejecución es incorrecta
    '''
    if dataset is None:
        print(u'\nFaltan argumentos por pasar a la función')
        return 1
    lista_variables_categoricas = []
    other = []
    for i in dataset.columns:
        if (dataset[i].dtype!=float) & (dataset[i].dtype!=int):
            unicos = int(len(np.unique(dataset[i].dropna(axis=0, how='all'))))
            if unicos < 100:
                lista_variables_categoricas.append(i)
            else:
                other.append(i)

    return lista_variables_categoricas, other

dame_variables_categoricas(cca_df)

No obstante, la mayoria de las variables de nuestro dataset son categóricas pero vienen ya codificadas. Chequear documento **drivingLegend.pdf** para referencia. El problema con muchas de estas variables es que son numericas pero no tienen un orden establecido. Es decir, que una variable tenga valores de 1,2,3... no quiere decir que 3 sea de mayor magnitud que 1. La variable en si no representa una relación lineal. Por tanto tendremos que hacer unas modificaciones.

# Creamos un csv con el dataframe completo 
cca_df.to_csv('data/cca_df.csv')
