==> 01_EDA_parte_1.ipynb <==
**19/12/2021**
<br>
Bruno Mazzilli González

> # EDA: parte 1

En este notebook se van preparar los dataframes para poder empezar a analizarlos

## Librerias

import numpy as np
import pandas as pd
#import matplotlib
#from matplotlib import pyplot as plt # plots
#import seaborn as sns # plots
#from sklearn.ensemble import IsolationForest
#from sklearn import preprocessing 
#from sklearn.preprocessing import MinMaxScaler

import warnings
warnings.filterwarnings("ignore")

## Importamos datos

data = pd.read_csv('data/cca_df.csv', low_memory=False) #low_memory = False es para evitar el warning por ser columnas de distinto formato

data.head(3)

<br>
Eliminamos la columna de índice 'Unnamed: 0', que se crea por defecto

# Al crear el data frame python exporta el csv añadiendo la columna del índice
del data['Unnamed: 0']

data.head(3)

data.shape

<br>

## ESTUDIO DE VARIABLES:
### Variable Target:
La variable objetivo del estudio que se quiere predecir es: ¿habrá fallecimientos o no? En el data set original esta información la proporciona la variables: "C_SEV"

A continuacion se analiza cual es la frecuencia de cada uno de los posibles valores de las variables "C_SEV" 

### Variable Target: Fallecimiento

data['C_SEV'].isna().value_counts()

data['C_SEV'].value_counts(dropna=False)

En primer lugar analizamos los accidentes que han resultado en algun fallecimiento

En función de los valores de la variable "C_SEV", se definen dos grupos:
Grupo 1: El accidente a resultado en almenos un fallecido
Grupo 2: El accidente a resultado en una lesion no mortal

Observamos que hay **98.633** accidentes con algun fallecimiento y **5.761.772** con alguna lesión no mortal.
Nuestros datos estan desbalanceados en la distribucion de los grupos 1 y 2. 
Para evitar que este 'bias' en el training set afecte a los algoritmos que usaremos tendremos que rebalancear los datos mas adelante

#creamos un dataframe especifico para esta variable target.
c_sev_df = data
# Creamos columna asociada con la variable target. toma valor 1 si ha habido algun fallecido y 0 en caso contrario
c_sev_df['target_c_sev'] = np.where((c_sev_df['C_SEV'] == 1.0), 1, 0)
#eliminamos la columna 'C_SEV' ya que ya tenemos una columna asociada con la variable target
del c_sev_df['C_SEV']
#excluimos la variable P_ID y V_ID ya que se considera el numero de identificacion de una persona no aporta informacion a 
#la variable target
del c_sev_df['P_ID']
del c_sev_df['V_ID']

print("Nuestro dataframe para el Target Fallecimiento tiene la forma:", c_sev_df.shape)

c_sev_df.head(3)

<br>

### Tratamiento de NAs

Ahora que ya tenemos el dataframe asociado a nuestra variable target procedemos a el tratamiento de valores missing

#creamos una funcion que identifique los valores NA en cada una de las variable de un dataframe

def identifica_NA(data):
    """
    Esta funcion devuelve un dataframe con el numero total de NAs y el porcentaje de NAs en cada columna.
    
    Parametros
    ----------
    data: dataframe
    """
    #creamos serie de pandas que cuente el numero de NAs para cada columna:
    null_sum = data.isnull().sum()
    #ponemos en orden de mas NAs a menos:
    total = null_sum.sort_values(ascending=False)
    #porcentaje de NAs
    percent = ( ((null_sum / len(data.index))*100).round(2) ).sort_values(ascending=False) #porcentaje de NAs
    
    # concatenamos columnas 'total' y 'percent' para crear el dataframe:
    df_NA = pd.concat([total, percent], axis=1, keys=['Number of NA', 'Percent NA'])
    # eliminamos filas que de este dataframe que no tengan valores NA. omitir si quieres saber que variables no tienen NAs
    #df_NA = df_NA[ (df_NA.T != 0).any() ]
    
    return df_NA

#### Tratamiento NAs dataframe con Variable Target: Fallecimiento

identifica_NA(c_sev_df)

Observando el numero de NAs en las columnas de nuestro dataframe podemos ver que no podemos simplemente eliminar todas las filas con NAs ya que esto eliminaría un gran porcentaje de datos. Por tanto eliminaremos solo los NAs que representan un porcentaje bajo del total de esa variable (hasta 5%). Los demás NAs los rellenaremos con la moda de esa columna.

#rellenamos los NAs con la media de su respectiva columna:
columnas_a_cambiar = ["P_SAFE", "C_RCFG", "V_YEAR", "C_CONF", "C_RALN", "P_AGE", "P_ISEV", "C_TRAF"]
for i in columnas_a_cambiar:
 c_sev_df[i] = c_sev_df[i].fillna( int(c_sev_df[i].mode() ))


identifica_NA(c_sev_df).head(7)

Eliminamos las demás filas con NAs:

c_sev_df =  c_sev_df.dropna()

#codificamos la variable 'P_SEX'
c_sev_df["P_SEX"] = c_sev_df["P_SEX"].replace({"F":"1", "M":"2"})

print("Nuestro dataframe para el Target Fallecimiento tiene la forma:", c_sev_df.shape)
c_sev_df.head(3)

c_sev_df.dtypes

#Nos aseguramos de que no hay mas valores NA:
identifica_NA(c_sev_df).head()

<br>

### Tratamiento de variables categoricas

#convertimos a strings variables categoricas codificadas como float. esto nos permitirá hacer el encoding. 
cols = ['C_VEHS','C_CONF','C_RCFG','C_WTHR','C_RSUR','C_RALN','C_TRAF','V_TYPE','P_PSN','P_ISEV','P_SAFE','P_USER']
for i in cols:
 c_sev_df[i] = c_sev_df[i].apply(str)

c_sev_df.dtypes

c_sev_df.head()

#dividimos el dataframe en variables categoricas y numericas
df_object = c_sev_df.select_dtypes(include='object')
df_numeric = c_sev_df.select_dtypes(exclude='object')

df_object.head()

df_numeric.head()

#hacemos el encoding. vamos a crear una nueva columna para cada posible valor en cada variable categórica
df_object_dummies = pd.get_dummies(df_object, drop_first=True)

df_object_dummies.shape

df_object_dummies.head(3)

Hay muchas columnas pero mas adelante haremos seleccion de variables para 'coger' la información importante

A continuación unimos los dataframes

final_df = pd.concat([df_object_dummies,df_numeric],axis=1)

final_df.head(3)

final_df.shape

final_df.dtypes.value_counts()

final_df.info()

final_df.info(verbose=True)

<br>

### Detección de Outliers

Se considera que no hay outliers para este dataset. Los outliers amenudo representan errores en el proceso de recoleccion de los datos. Sin embargo los datos de nuestro dataset representan variables categóricas codificadas y por tanto los valores que pueden tomar estas variables estan limitadas a un rango establecido. En conclusión, es muy poco probable que los valores de nuestras variables representen valores atípicos.

<br>

#### Una vez ya tenemos los dataframes preparados procedemos a cargarlos a un csv para usarlos mas adelante:

#guardamos el final_df como parquet para facilitar el procesamiento
c_sev_df.to_csv('data/c_sev_df.csv')
final_df.to_parquet('data/c_sev_df_encoded.parquet')
